# docs/swagger.yaml
definitions:
  datastore.Admin:
    properties:
      created_tags:
        items:
          $ref: '#/definitions/datastore.Tag'
        type: array
      createdAt:
        description: add ``
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      password:
        type: string
      updatedAt:
        description: add ``
        type: string
    type: object
  datastore.Card:
    properties:
      approved:
        type: string
      company_id:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      linkToProd:
        type: string
      name:
        type: string
      price:
        type: string
      tags:
        items:
          type: string
        type: array
      views:
        type: integer
    type: object
...
  datastore.Tag:
    properties:
      admin_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  utility.Message:
    properties:
      text:
        type: string
    type: object
host: http://localhost:8081
info:
  contact: {}
  description: Nenavijy swagger
  license:
    name: BSD-3
  title: pay-with-crypto API
  version: "1.0"
paths:
  /admin/getForApprove:
    get:
      consumes:
      - application/json
      description: Get cards for validate
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/datastore.Card'
            type: array
      security:
      - accessToken: []
      tags:
      - Admin
  /auth/admin_login:
    post:
      consumes:
      - application/json
      description: Login to admin account
      parameters:
      - description: Admin data
        in: body
        name: admin_data
        required: true
        schema:
          properties:
            name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datastore.Admin'
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Admin already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Auth
  /auth/admin_register:
    post:
      consumes:
      - application/json
      description: Create new admin accounts tokens.
      parameters:
      - description: Admin data
        in: body
        name: admin_data
        required: true
        schema:
          properties:
            first_name:
              type: string
            last_name:
              type: string
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datastore.Admin'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utility.Message'
        "409":
          description: Admin already created
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      security:
      - accessToken: []
      tags:
      - Auth
  /auth/createTag:
    post:
      consumes:
      - application/json
      description: Login to admin account
      parameters:
      - description: Tag data
        in: body
        name: tag_data
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/datastore.Tag'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Admin
...
  /auth/softDelete:
    delete:
      consumes:
      - application/json
      description: Ban company account
      parameters:
      - description: Company data
        in: body
        name: company_data
        required: true
        schema:
          properties:
            id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.Message'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utility.Message'
        "404":
          description: Company not exist
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Admin
...
  /auth/validateCard:
    put:
      consumes:
      - application/json
      description: Validate company card as admin
      parameters:
      - description: Validate data
        in: body
        name: validate_data
        required: true
        schema:
          properties:
            id:
              type: string
            status:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.Message'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utility.Message'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utility.Message'
      tags:
      - Admin
...
securityDefinitions:
  string:
    in: header
    name: accessToken
    type: apiKey
swagger: "2.0"